Bridge backend to frontend: fix data mismatches, enforce ADMIN, add error handling
Align the API with the Robotbino/gameStore React frontend and fix several
latent bugs that broke key flows:

- Accept genre as a JSON array on game create/update (GenreDeserializer),
  matching what the frontend sends; still stored comma-separated
- JWT tokens now honor jwt.expiration (24h) instead of a hardcoded 24 min
- JWTAuthenticationFilter always continues the chain; invalid/expired tokens
  no longer hang requests, they proceed unauthenticated -> clean 401
- UsersService: fix updateUser (correct userName getter, keep existing
  password when none supplied), encode passwords and reject duplicates in
  addUser; UsersController returns the saved entity and is now @RestController
- Enforce role-based access: GET /games public, all other game/user
  management is ADMIN-only; return 401 (not 403) when unauthenticated so the
  frontend redirects to login; permit /error to preserve 403s
- Hide password and internal UserDetails fields from JSON responses; expose
  userName (Lombok skipped it due to the getUsername() override)
- Global exception handler: 404 for missing records, 401 for bad
  credentials, 400 for other errors, all as {"message": ...}
- Align Users id / repository / controller on Integer to match the DB schema
- Update README (access_token, ADMIN access, error responses, roadmap)

+++ src/main/java/com/gameStore/Bino/repositories/UserRepository.java
@@ -1,14 +1,12 @@
 package com.gameStore.Bino.repositories;
 
-import com.gameStore.Bino.models.Games;
 import com.gameStore.Bino.models.Users;
-import org.apache.catalina.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<Users,Long>  {
+public interface UserRepository extends JpaRepository<Users,Integer>  {
 
     Optional<Users> findByEmail(String email);
